+++ src/main/kotlin/com/michalpu/PokeRest/client/PokemonClient.kt
@@ -1,7 +1,5 @@
 package com.michalpu.PokeRest.client
 
-import com.michalpu.PokeRest.domain.Pokemon
-
 interface PokemonClient {
-    fun fetchByName(name: String): Pokemon?
+    fun getByName(name: String): Pokemon?
 }
