+++ src/main/kotlin/com/michalpu/PokeRest/client/PokemonClientException.kt
@@ -1,0 +1,3 @@
+package com.michalpu.PokeRest.client
+
+class PokemonClientException(cause: Throwable) : RuntimeException("Pokemon client exception", cause)
