--- src/main/kotlin/com/michalpu/PokeRest/domain/Greeting.kt
@@ -1,4 +1,0 @@
-package com.michalpu.PokeRest.domain
-
-data class Greeting (val id: Long, val content: String){
-}
