+++ src/main/kotlin/com/michalpu/PokeRest/pokemon/PokemonApiRestTemplateConfiguration.kt
@@ -1,4 +1,4 @@
-package com.michalpu.PokeRest.config
+package com.michalpu.PokeRest.pokemon
 
 import org.apache.http.impl.client.HttpClientBuilder
 import org.springframework.beans.factory.annotation.Value
@@ -6,7 +6,6 @@ import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.http.client.HttpComponentsClientHttpRequestFactory
 import org.springframework.web.client.RestTemplate
-import kotlin.math.max
 
 @Configuration
 class PokemonApiRestTemplateConfiguration {
