--- src/main/kotlin/com/michalpu/PokeRest/service/PokemonService.kt
@@ -1,7 +1,0 @@
-package com.michalpu.PokeRest.service
-
-import com.michalpu.PokeRest.domain.Pokemon
-
-interface PokemonService {
-    fun getByName(name: String) : Pokemon?
-}
