+++ src/test/kotlin/com/michalpu/PokeRest/client/PokemonClientImplTest.kt
@@ -1,23 +1,18 @@
 package com.michalpu.PokeRest.client
 
 import com.michalpu.PokeRest.PokeRestApplicationTests
-import com.michalpu.PokeRest.config.SpyRestTemplate
-import com.michalpu.PokeRest.domain.Pokemon
 import org.junit.jupiter.api.Test
 
 class PokemonClientImplTest() : PokeRestApplicationTests() {
 
-    lateinit var pokemonClient: PokemonClientImpl
-    lateinit var spyRestTemplate: SpyRestTemplate
+    var spyRestTemplate = SpyRestTemplate()
+    var pokemonClient: PokemonClientImpl = PokemonClientImpl(spyRestTemplate, "https://pokeapi.co/api/v2")
 
 
     @Test
     fun shouldReturnAPokemon (){
 
         //given
-        spyRestTemplate = SpyRestTemplate()
-        pokemonClient = PokemonClientImpl(spyRestTemplate, "https://pokeapi.co/api/v2")
-
         spyRestTemplate.stubbedGet = """
             {
                 "id" : 4,
@@ -27,7 +22,7 @@ class PokemonClientImplTest() : PokeRestApplicationTests() {
         """.trimIndent()
 
         //when
-        val pokemon = pokemonClient.fetchByName("charmander")
+        val pokemon = pokemonClient.getByName("charmander")
 
         //then
         assert(pokemon != null)
