+++ src/test/kotlin/com/michalpu/PokeRest/client/SpyRestTemplate.kt
@@ -1,4 +1,4 @@
-package com.michalpu.PokeRest.config
+package com.michalpu.PokeRest.client
 
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.KotlinModule
